pyproject.toml
[tool.poetry]
name = "gitsec"
version = "1.0.1"
description = "Python tool that helps you enhance your GitHub security posture at scale."
authors = ["Aplite hi@aplite.de"]
license = "MIT"
readme = "README.md"
packages = [{ include = "gitsec" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.20"
pydantic = "2.12.3"
httpx = "0.28.1"
detect-secrets = "1.5.0"
gitpython = "3.1.45"
openpyxl = "3.1.5"

[tool.poetry.scripts]
gitsec = "gitsec.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
